refactor: split large files, Cards page, lint fix
- CardModal: extract CardModalMoveContent, CardModalBackgroundPicker
- CardsTable: extract types, utils, CardsTableToolbar
- ListColumn: extract ListColumnDialogs
- BoardMenu: extract BoardMenuDialogs
- eventHandlers: split into listEventHandlers, cardEventHandlers
- Topbar: Cards link to /cards
- Sidebar: Cards entry per workspace
- Fix lint: remove unused targetList in cardEventHandlers
