+++ frontend/app/boards/[id]/components/BoardHeader/BoardMenu.tsx
@@ -30,22 +30,6 @@ import {
   DropdownMenuSubTrigger,
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
 import { getVisibilityLabel, getVisibilityIcon } from './utils';
 import { getVisibilityDescription } from '@/components/BoardView';
@@ -66,6 +50,7 @@ import {
   ItemDescription,
 } from '@/components/ui/item';
 import { ArchivedItemsContent } from './ArchivedItemsDialog';
+import { BoardMenuDialogs } from './BoardMenuDialogs';
 import type { BoardMember } from '../../types';
 import type { Board } from '../../types';
 import type { List } from '../../types';
@@ -501,143 +486,27 @@ export function BoardMenu({
         </DropdownMenuContent>
       </DropdownMenu>
 
-      {/* Copy Board Dialog */}
-      <Dialog open={showCopyDialog} onOpenChange={setShowCopyDialog}>
-        <DialogContent className='sm:max-w-md border-accent'>
-          <DialogHeader>
-            <DialogTitle>Copy board</DialogTitle>
-            <DialogDescription>
-              Create a new board with the same title, description, background
-              and visibility. The new board will have default lists (To Do,
-              Doing, Done).
-            </DialogDescription>
-          </DialogHeader>
-          <div className='space-y-4 py-4'>
-            <div className='space-y-2'>
-              <Label htmlFor='copy-board-title'>Board name</Label>
-              <Input
-                id='copy-board-title'
-                placeholder='Enter board name'
-                value={copyBoardTitle}
-                onChange={(e) => setCopyBoardTitle(e.target.value)}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') {
-                    handleCopyBoard();
-                  }
-                }}
-              />
-            </div>
-            <div className='flex justify-end gap-2'>
-              <Button
-                variant='outline'
-                onClick={() => setShowCopyDialog(false)}
-              >
-                Cancel
-              </Button>
-              <Button
-                onClick={handleCopyBoard}
-                disabled={copying || !copyBoardTitle.trim()}
-              >
-                {copying ? 'Copying...' : 'Copy board'}
-              </Button>
-            </div>
-          </div>
-        </DialogContent>
-      </Dialog>
-
-      {/* Leave board (assign admin first) Dialog */}
-      <Dialog open={showLeaveBoardDialog} onOpenChange={setShowLeaveBoardDialog}>
-        <DialogContent className='sm:max-w-md border-accent'>
-          <DialogHeader>
-            <DialogTitle>Assign an admin before leaving</DialogTitle>
-            <DialogDescription>
-              You are the last admin on this board. Choose a member to assign as
-              admin, then you can leave the board.
-            </DialogDescription>
-          </DialogHeader>
-          <div className='space-y-4 py-4'>
-            <div className='space-y-2'>
-              <Label>Member to assign as admin</Label>
-              <Select
-                value={leaveBoardAdminUserId}
-                onValueChange={setLeaveBoardAdminUserId}
-              >
-                <SelectTrigger className='w-full'>
-                  <SelectValue placeholder='Select a member...' />
-                </SelectTrigger>
-                <SelectContent>
-                  {otherBoardMembersToPromote.map((m) => (
-                    <SelectItem key={m.userId} value={m.userId}>
-                      {m.user?.name || m.user?.email || 'Member'}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-            </div>
-            <div className='flex justify-end gap-2'>
-              <Button
-                variant='outline'
-                onClick={() => setShowLeaveBoardDialog(false)}
-              >
-                Cancel
-              </Button>
-              <Button
-                onClick={handleAssignAdminAndLeaveBoard}
-                disabled={leavingBoard || !leaveBoardAdminUserId}
-                className='bg-orange-500 hover:bg-orange-600 text-white border-0'
-              >
-                {leavingBoard ? 'Leaving...' : 'Assign as admin and leave'}
-              </Button>
-            </div>
-          </div>
-        </DialogContent>
-      </Dialog>
-
-      {/* Email-to-board Dialog */}
-      <Dialog open={showEmailDialog} onOpenChange={setShowEmailDialog}>
-        <DialogContent className='sm:max-w-md border-accent'>
-          <DialogHeader>
-            <DialogTitle>Email-to-board</DialogTitle>
-            <DialogDescription>
-              Send emails to this board. Emails will be converted to cards.
-            </DialogDescription>
-          </DialogHeader>
-          <div className='space-y-4 py-4'>
-            <div className='space-y-2'>
-              <Label htmlFor='email-address'>Email address</Label>
-              <Input
-                id='email-address'
-                type='email'
-                placeholder='Enter email address'
-                value={emailAddress}
-                onChange={(e) => setEmailAddress(e.target.value)}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') {
-                    handleEmailToBoard();
-                  }
-                }}
-              />
-              <p className='text-xs text-muted-foreground'>
-                Send emails to this address to create cards on this board
-              </p>
-            </div>
-            <div className='flex justify-end gap-2'>
-              <Button
-                variant='outline'
-                onClick={() => setShowEmailDialog(false)}
-              >
-                Cancel
-              </Button>
-              <Button
-                onClick={handleEmailToBoard}
-                disabled={sendingEmail || !emailAddress.trim()}
-              >
-                {sendingEmail ? 'Sending...' : 'Save'}
-              </Button>
-            </div>
-          </div>
-        </DialogContent>
-      </Dialog>
+      <BoardMenuDialogs
+        showCopyDialog={showCopyDialog}
+        setShowCopyDialog={setShowCopyDialog}
+        copyBoardTitle={copyBoardTitle}
+        setCopyBoardTitle={setCopyBoardTitle}
+        copying={copying}
+        onCopyBoard={handleCopyBoard}
+        showEmailDialog={showEmailDialog}
+        setShowEmailDialog={setShowEmailDialog}
+        emailAddress={emailAddress}
+        setEmailAddress={setEmailAddress}
+        sendingEmail={sendingEmail}
+        onEmailToBoard={handleEmailToBoard}
+        showLeaveBoardDialog={showLeaveBoardDialog}
+        setShowLeaveBoardDialog={setShowLeaveBoardDialog}
+        leaveBoardAdminUserId={leaveBoardAdminUserId}
+        setLeaveBoardAdminUserId={setLeaveBoardAdminUserId}
+        leavingBoard={leavingBoard}
+        otherBoardMembersToPromote={otherBoardMembersToPromote}
+        onAssignAdminAndLeave={handleAssignAdminAndLeaveBoard}
+      />
 
       <ChangeBackgroundDialog
         open={showBackgroundDialog}
